feat(bump): add "allowed-branches" and "create-tag" support
"allowed-branches", configurable in the config file (default ".*"),
is a regex that will only allow releases/tags to be made when the
context's ref matches the regex.

"create-tag" is a new Action input, which, when set to "true", will
create a lightweight Git tag only, as opposed to "create-release"'s
full GitHub release.