fix(bump): switch to GraphQL query to fetch tags by date
The REST API call used previously is both not properly documented and
did not behave as assumed, leading to failure to resolve the latest
tags if the version scheme changed and the total amount of tags in the
repo exceeds 100.

We now order explicitly by GitHub's own `TAG_COMMIT_DATE`, which
introduces another caveat; the documentation for that order method says:

 > Order refs by underlying commit date if the ref prefix is refs/tags/

No mention of how that is implemented or what Git data it uses under the
hood. The phrasing _hints_ that it uses the ref's `committerdate`, which
would be invalid for annotated tags, but during testing, it appears to
"work fine", which means it probably uses `creatordate` or (more likely)
a custom implementation.