improvement(C019): consider ticket whitespace boundary
This takes a bit more of the context of the string into account when
checking for ticket-like strings in the subject line.
A trailing `-` character no longer triggers the rule, so that (e.g.)
CVE issues with the format `CVE-1234-56789` stop triggering it. (And
`CVE-` was added to the allowed words list for good measure.)

Also, rewrite the check to be more efficient; all checks could be moved
to the regex itself. And add fancy column number `LlvmRange`.