improvement(sdkver): make changelog relative to last full release
Instead of the "current version", which may be an -rc or -dev version,
we now always get the list of commits since the last full release and
base our changelog on that.
A "full release" is any release without a prerelease marker '-'.
Changelog generation has seen some refactoring to accommodate that,
as the described flow requires different/customizable source data.