chore: change `revert` type tag to bump patch by default
A `revert` commit is a non-standard occurrence; our own experience has
taught us that it is most likely to happen under time pressure to
circumvent a change in the field. Without such time pressure, a `fix`
commit is more likely to be used.

In other words, when we encounter a `revert` commit, quick deployment
is likely of the essence, which often happens only when a bump is
performed.
So to ensure a patch bump, users need to have commit message like:
  fix: revert change to X

However, people tend to use:
  revert: change to X

The latter is recommended by the Conventional Commits spec page:
  https://www.conventionalcommits.org/en/v1.0.0/#how-does-conventional-commits-handle-revert-commits

When such a `revert` type lands in the mainline, and the code is not
deployed since no bump is performed, people need to follow up with a
dummy `fix` PR and commit, wasting time and resources.

So, let's slightly deviate from spec by default and just bump `PATCH`
on `revert` type commits, and risk bumping patch more often than one
would expect.