feat: support GitHub cutting-lines in git-trailers
This commit introduces support for cutting-lines, as inserted
by GitHub in certain merge strategies. In example:

```
feat: remove logout button

BREAKING-CHANGE: this removes the logout button as we want
  all users to be trapped in this application

--------

co-authored-by: Bob the Builder
```

This would originally result in only the `co-authored-by` trailer
to be accepted, dropping the `BREAKING CHANGE` flag (and therefor
only bumping the MINOR version).

This change will ignore the cutting-line and consider both elements
in the above example as valid trailers.

NOTE: the cutting-line is the only exception, any other paragraph
      entry will still discard previous git-trailers; i.e.

```
feat: remove logout button

BREAKING-CHANGE: this removes the logout button as we want
  all users to be trapped in this application

--------

This parapgraph causes the BREAKING-CHANGE to be discarded

co-authored-by: Bob the Builder
```