fix(bump): paginate and get all tags while determining bump
We used to fetch only the 100 latest tags; that is not sufficient in
a repository with heavy traffic, especially when an older release
requires a fix up.

The way we currently determine the last matching tags unfortunately
requires us to fetch either all commits or all tags in a repository.