fix: handle breaking changes when current version is dev
This commit restructures the SdkVer handling based on use-case
mapping, simplifying the overall complexity of maintaining the
flow.

This was needed to tackle the specific use case for failure of
creating a new breaking change from a `dev` or `rc` version.

i.e. `1.0.0-dev001.SHA` could not lead to `2.0.0-rc01` even
     when specifying that this is a breaking change.