fix(bump): remove non-compliance reporting from SdkVer
For SdkVer, we can be much more tolerant of Conventional Commit
noncompliance, as there's very little functional ties to SdkVer.
So `warning`-level really isn't warranted there.
In fact, removing the check altogether can be seen as an improvement,
as in the context of a "flexible" commit description repository, the
noncompliance report really only establishes an unhealthy level of
"baseline warning" spam.